+++ CHANGES.rst
@@ -101,3 +101,4 @@ v1.0.11
 
 v1.0.12[unreleased]
 -----------
+* Add admin email