+++ CHANGES.rst
@@ -105,3 +105,4 @@ v1.0.12[unreleased]
 * Add causes to Organization serializers
 * Add 'contact_name', 'contact_phone' and 'contact_email' to Organization model
 * Add owner to OrganizationRetrieveSerializer
+* Add pt_BR translations